EPI-644: Build reusable scaffold for data migrations
