+++ packages/sync-server/__tests__/subCommands/migrateChangelogNotesToEncounterHistory.test.js
@@ -13,7 +13,7 @@ import {
 import { sleepAsync } from '@tamanu/shared/utils';
 
 import { createTestContext } from '../utilities';
-import { migrateChangelogNotesToEncounterHistory } from '../../app/subCommands';
+import { migrateData } from '../../app/subCommands';
 
 const DEFAULT_USER_ID = 'DEFAULT_USER_ID';
 
@@ -129,9 +129,8 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
   let locationGroup2;
   let defaultUser;
 
-  const SUB_COMMAND_OPTIONS = {
-    batchSize: 1,
-  };
+  const SUB_COMMAND_NAME = 'ChangelogNotesToEncounterHistory';
+  const SUB_COMMAND_OPTIONS = { batchSize: 1, delay: 0 };
 
   const getDateSubtractedFromNow = daysToSubtract =>
     toDateTimeString(sub(new Date(), { days: daysToSubtract }));
@@ -261,7 +260,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.locationId = newLocation.id;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -313,7 +312,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.departmentId = newDepartment.id;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -365,7 +364,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.examinerId = newClinician.id;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -416,7 +415,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.encounterType = 'clinic';
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -469,7 +468,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.encounterType = 'clinic';
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -567,7 +566,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.encounterType = newEncounterType;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -683,7 +682,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.locationId = location4.id;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -787,7 +786,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.departmentId = department4.id;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -891,7 +890,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.examinerId = clinician4.id;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -995,7 +994,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.encounterType = encounterType4;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -1154,7 +1153,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.encounterType = encounterType4;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -1311,7 +1310,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       encounter.encounterType = encounterType4;
       await encounter.save();
 
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -1420,7 +1419,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await encounter.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -1535,7 +1534,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await encounter.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -1641,7 +1640,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         encounter.locationId = location4.id;
         await encounter.save();
 
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -1734,7 +1733,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         encounter.locationId = location4.id;
         await encounter.save();
 
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -1829,7 +1828,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         encounter.locationId = location4.id;
         await encounter.save();
 
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -1922,7 +1921,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         location2.name = 'Changed location 2';
         await location2.save();
 
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -1998,7 +1997,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         encounter.departmentId = department3.id;
         await encounter.save();
 
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2095,7 +2094,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await location3.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2191,7 +2190,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await encounter.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2286,7 +2285,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await encounter.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2381,7 +2380,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await encounter.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2478,7 +2477,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await encounter.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2572,7 +2571,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await department2.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2654,7 +2653,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         encounter.locationId = location3.id;
         await encounter.save();
 
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2740,7 +2739,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await encounter.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2827,7 +2826,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         await clinician2.save();
 
         // Migration
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2907,7 +2906,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
         encounter.locationId = location3.id;
         await encounter.save();
 
-        await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+        await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
         expect(exitSpy).toBeCalledWith(0);
 
@@ -2974,7 +2973,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       );
 
       // Migration
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -3037,7 +3036,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       });
 
       // Migration
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -3122,10 +3121,10 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       });
 
       // Migration 1
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       // Migration 2
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -3197,7 +3196,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       );
 
       // Migration
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
 
@@ -3248,7 +3247,7 @@ describe('migrateChangelogNotesToEncounterHistory', () => {
       );
 
       // Migration
-      await migrateChangelogNotesToEncounterHistory(SUB_COMMAND_OPTIONS);
+      await migrateData(SUB_COMMAND_NAME, SUB_COMMAND_OPTIONS);
 
       expect(exitSpy).toBeCalledWith(0);
