+++ packages/sync-server/app/subCommands/index.js
@@ -2,13 +2,17 @@ export * from './apiKeys';
 export * from './calculateSurveyResults';
 export * from './fhir';
 export * from './generate';
+export * from './generateInitialVitalLogs';
 export * from './generateVaccineCertificate';
 export * from './importReport';
 export * from './loadSigner';
 export * from './migrate';
+export * from './migrateData';
 export * from './migrateImagingRequestsToLocationGroups';
+export * from './migrateVitals';
 export * from './printRoutes';
 export * from './provision';
+export * from './removeDuplicatedDischarges';
 export * from './report';
 export * from './saveCertificateRequest';
 export * from './serve';
@@ -17,8 +21,3 @@ export * from './settings';
 export * from './shell';
 export * from './tasks';
 export * from './user';
-export * from './migrateVitals';
-export * from './generateInitialVitalLogs';
-export * from './migrateChangelogNotesToEncounterHistory';
-export * from './removeDuplicatedDischarges';
-export * from './migrateNotePagesToNotes';
