+++ packages/sync-server/app/subCommands/migrateData/dataMigrations/index.js
@@ -1,0 +1,2 @@
+export * from './ChangelogNotesToEncounterHistory';
+export * from './NotePagesToNotes';
