+++ packages/sync-server/app/subCommands/migrateData/index.js
@@ -1,0 +1,1 @@
+export { migrateDataCommand } from './migrateData';
